+++ CHANGELOG.md
@@ -7,6 +7,7 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 ## Unreleased
 
 - Support making fields infallible with `@juniper(infallible: true)`.
+- Make sure fields on input object structs are in fact public.
 
 #### Breaking changes